fix the bug in offline barrel write.
That is, in writing temp file for 2nd barrel, that temp file is already closed by prefious flush, while it's not opened again.

This bug was introduced by previous bug fix for online barrel write,
in that case, IndexBarrelWriter is not deleted in IndexWriter::flush() in case of other query thread would still use IndexBarrelWriter.

In the new fix, FieldIndexer::reset() would reopen the temp barrel file.