refactor IndexMerger and its subclasses.
1. In original design, BTMerger, OptimizeMerger, etc inherits from
IndexMerger, which makes it hard to reuse the subclass functions if we
want to change the behavior of IndexMerger (fox example, create a
MockIndexMerger). So we use composition relationship instead of inheritance.

2. The new class "MockIndexMerger" replaces original "MockMerger",
it simulates barrels merge while not loading and writing barrel files.

3. The new test suite "t_IndexMergePolicy" replaces original
"t_mergepolicy", it uses "MockIndexMerger" to test "BTMerger" and
"OptimizeMerger".

4. IndexMerger::mergeBarrel() is too long, so it is refactored into two
functions "outputNewBarrel()" and "createNewBarrelInfo()".