fix bugs in MultiTermReader and BarrelInfo.
1. in MultiTermReader::loadReader(), if exception is thrown in
TermReader::clone(), the newed ReaderCache() might not be deleted,
so use auto_ptr to store the newed ReaderCache() and release it
before insert it into readerCache_.

2. when merge thread is removing a barrel, BarrelInfo::setDirty()
is called, it iterates the set of indexInputs.
If query thread is deleting TermDocFreqs instance,
BarrelInfo::unRegisterIndexInput() is called, it would erase an element
in indexInputs, so that the iterator in merge thread is invalid,
which might cause memory access fault.
So a lock is added in BarrelInfo::setDirty() to avoid execute
BarrelInfo::unRegisterIndexInput() concurrently.