fix the bug of non-atomicity of BarrelsInfo in IndexMerger::createNewBarrelInfo().
In IndexMerger::createNewBarrelInfo(), it iterates the barrels to remove.
When a barrel is removed, and before removing other barrels and adding the new merged barrel,
if BarrelsInfo::getDocCount() is called, it would return incorrect doc count,
that is, short of the doc count in the just removed barrel.

Although the lock is put in the functions of BarrelsInfo::removeBarrel()
and BarrelsInfo::addBarrel(), the sequence calls of these functions are
not atomic, which might lead to inconsistent barrels status.

This bug is fixed by moving the lock outside into the scope of IndexMerger::createNewBarrelInfo(),
so that the sequence calls of BarrelsInfo functions are atomic.