to enable query in multi-thread, remove TermReader::pFieldInfo_.
as FieldInfo is owned by IndexBarrelReader, and it is deleted when some
thread calls IndexReader::reopen(), so that if TermReader maintains the
reference to FieldInfo, memory access fault might happen.

to avoid this problem, make SparseTermReaderImpl maintain a copy of FieldInfo,
so that it could be shared by multiple TermReader instances.