fix the memory leak when exception is thrown in PostingReader constructor.
as the PostingReaders such as RTPostingReader owns its resources such as
InputDescriptor and SkipListReader by pointer, if any exception is
thrown in constructor, these resouces would be leaked.

to avoid this problem, declare these resources in boost::scoped_ptr objects,
so that these objects which have been constructed would be destroyed
automatically whenever exception is thrown in constructor.