Merge commit 54 of leveldb
- Replace raw slice comparison with a call to user comparator.

- Fix end of namespace comments.

- Fixed bug in picking inputs for a level-0 compaction.

When finding overlapping files, the covered range may expand
as files are added to the input set.  We now correctly expand
the range when this happens instead of continuing to use the
old range.  For example, suppose L0 contains files with the
following ranges:

        F1: a .. d
        F2:    c .. g
        F3:       f .. j

and the initial compaction target is F3.  We used to search
for range f..j which yielded {F2,F3}.  However we now expand
the range as soon as another file is added.  In this case,
when F2 is added, we expand the range to c..j and restart the
search.  That picks up file F1 as well.

This change fixes a bug related to deleted keys showing up
incorrectly after a compaction as described in Issue 44.

(Sync with upstream @25072954)