Fix bug on Sf1Watcher and ZooKeeperTopology
Tests showed that the zookeeper watcher mechanism does not properly respond to
node deletion events. It happens that the SF1 restart (that is a start after a
stop) is not correctly detected. This is because zookeeper watches are only
one-time triggers. Hence it is required to properly re-set watches after a
node deletion detection.

List of moidifications:
- added watch on node deletion event handler
- renamed constants in ZooKeeperNamespace header for better understanding
- renamed private methods in ZooKeeperRouter for avoinding confusion