Fix connection release problem
Do not cache Boost.Asio objects for network communication
because in case of connection lost such object are invalid.
That is, asking for a reconnection cause an exception to be
thrown (Broken pipe, Endpoint not valid).
Moreover, the connection pool must try to re-connect on connection
aquisition instead of on connection release.
Last, no exception specification shall be used, because the
run-time type checking that the C++ compiler performs can cause
an undeclared exception to abort the execution.