Fixed Bug 83
A retry mechanism has been added to the Sf1Driver class.
Only after using a broken connection it is possible to identify it as such,
so a single retry is performed. If the connection is still broken (e.g. because
the remote endpoint is not available) the current request fails.